Update chrono crate to fix RUSTSEC-2020-0159
See <https://rustsec.org/advisories/RUSTSEC-2020-0159> for more
information on the vulnerability, a potential segmenatation fault
in `localtime_r` invocations.

There are even newer versions of chrono than the one used in this
commit, but they would either pull in more dependencies or they
would not compile with rustc 1.56.0 which seems to be the current
MSRV for this project.